Add new answer to a challenge
Adding new challenge of easy category. New images are uploaded with the statement and the output in the images folder. Also the 'All solutions.sql' file in the 'Easy' folder is updated and a new md file is created with the statement, the code with the answer and its explanations.